+++ src/gui/mod.rs
@@ -1,2 +1,3 @@
 mod gui;
 pub use gui::*;
+mod menu;
